test:improve unit test
