+++ src/test/java/org/tron/common/storage/leveldb/LevelDbDataSourceImplTest.java
@@ -1,16 +1,19 @@
 /*
- * Copyright (c) [2016] [ <ether.camp> ] This file is part of the ethereumJ library.
+ * Copyright (c) [2016] [ <ether.camp> ]
+ * This file is part of the ethereumJ library.
  *
- * The ethereumJ library is free software: you can redistribute it and/or modify it under the terms
- * of the GNU Lesser General Public License as published by the Free Software Foundation, either
- * version 3 of the License, or (at your option) any later version.
+ * The ethereumJ library is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
  *
- * The ethereumJ library is distributed in the hope that it will be useful, but WITHOUT ANY
- * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
- * PURPOSE. See the GNU Lesser General Public License for more details.
+ * The ethereumJ library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details.
  *
- * You should have received a copy of the GNU Lesser General Public License along with the ethereumJ
- * library. If not, see <http://www.gnu.org/licenses/>.
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with the ethereumJ library. If not, see <http://www.gnu.org/licenses/>.
  */
 
 package org.tron.common.storage.leveldb;
@@ -18,9 +21,14 @@ package org.tron.common.storage.leveldb;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
 
+import java.util.HashMap;
+import java.util.Map;
 import org.junit.Before;
 import org.junit.Test;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.tron.common.utils.ByteArray;
 import org.tron.core.Constant;
 import org.tron.core.config.Configuration;
@@ -28,59 +36,113 @@ import org.tron.core.config.args.Args;
 
 public class LevelDbDataSourceImplTest {
 
-  private LevelDbDataSourceImpl dataSource;
+  private static final Logger logger = LoggerFactory.getLogger("Test");
+
+  LevelDbDataSourceImpl dataSourceTest;
 
   @Before
-  public void setup() {
+  public void initDb() {
     Args.setParam(new String[]{}, Configuration.getByPath(Constant.TEST_CONF));
-    dataSource = new LevelDbDataSourceImpl(Args.getInstance().getOutputDirectory(), "test");
+    dataSourceTest = new LevelDbDataSourceImpl(Args.getInstance().getOutputDirectory(), "test");
   }
 
   @Test
-  public void testGet() {
-    dataSource.initDB();
-    String key1 = "000134yyyhy";
+  public void testPutGet() {
+    dataSourceTest.resetDb();
+    String key1 = "2c0937534dd1b3832d05d865e8e6f2bf23218300b33a992740d45ccab7d4f519";
     byte[] key = key1.getBytes();
+    dataSourceTest.initDB();
+    String value1 = "50000";
+    byte[] value = value1.getBytes();
+
+    dataSourceTest.putData(key, value);
+
+    assertNotNull(dataSourceTest.getData(key));
+    assertEquals(1, dataSourceTest.allKeys().size());
+    assertEquals("50000", ByteArray.toStr(dataSourceTest.getData(key1.getBytes())));
+
+  }
+
+  @Test
+  public void testReset() {
+    LevelDbDataSourceImpl dataSource = new LevelDbDataSourceImpl(
+        Args.getInstance().getOutputDirectory(), "test_reset");
+    dataSource.resetDb();
+    assertEquals(0, dataSource.allKeys().size());
+    dataSource.closeDB();
+  }
+
+  @Test
+  public void testupdateByBatchInner() {
+    LevelDbDataSourceImpl dataSource = new LevelDbDataSourceImpl(
+        Args.getInstance().getOutputDirectory(), "test_updateByBatch");
+    dataSource.initDB();
+    dataSource.resetDb();
+    String key1 = "431cd8c8d5abe5cb5944b0889b32482d85772fbb98987b10fbb7f17110757350";
+    String value1 = "50000";
+    String key2 = "431cd8c8d5abe5cb5944b0889b32482d85772fbb98987b10fbb7f17110757351";
+    String value2 = "10000";
+
+    Map<byte[], byte[]> rows = new HashMap<>();
+    rows.put(key1.getBytes(), value1.getBytes());
+    rows.put(key2.getBytes(), value2.getBytes());
+
+    dataSource.updateByBatch(rows);
+
+    assertEquals("50000", ByteArray.toStr(dataSource.getData(key1.getBytes())));
+    assertEquals("10000", ByteArray.toStr(dataSource.getData(key2.getBytes())));
+    assertEquals(2, dataSource.allKeys().size());
+  }
+
+  @Test
+  public void testdeleteData() {
+    LevelDbDataSourceImpl dataSource = new LevelDbDataSourceImpl(
+        Args.getInstance().getOutputDirectory(), "test_delete");
+    dataSource.initDB();
+    String key1 = "431cd8c8d5abe5cb5944b0889b32482d85772fbb98987b10fbb7f17110757350";
+    byte[] key = key1.getBytes();
+    dataSource.deleteData(key);
     byte[] value = dataSource.getData(key);
     String s = ByteArray.toStr(value);
-    dataSource.closeDB();
-    System.out.println(s);
+    assertNull(s);
+
   }
 
   @Test
-  public void testPut() {
+  public void testallKeys() {
+    LevelDbDataSourceImpl dataSource = new LevelDbDataSourceImpl(
+        Args.getInstance().getOutputDirectory(), "test_find_key");
     dataSource.initDB();
-    String key1 = "000134yyyhy";
+    dataSource.resetDb();
+
+    String key1 = "431cd8c8d5abe5cb5944b0889b32482d85772fbb98987b10fbb7f17110757321";
     byte[] key = key1.getBytes();
 
     String value1 = "50000";
     byte[] value = value1.getBytes();
 
     dataSource.putData(key, value);
+    String key3 = "431cd8c8d5abe5cb5944b0889b32482d85772fbb98987b10fbb7f17110757091";
+    byte[] key2 = key3.getBytes();
 
-    assertNotNull(dataSource.getData(key));
-    assertEquals(1, dataSource.allKeys().size());
+    String value3 = "30000";
+    byte[] value2 = value3.getBytes();
 
-    dataSource.closeDB();
-  }
-
-  @Test
-  public void testRest() {
+    dataSource.putData(key2, value2);
+    assertEquals(2, dataSource.allKeys().size());
     dataSource.resetDb();
-    dataSource.closeDB();
   }
 
   @Test(timeout = 100)
   public void testLockReleased() {
-    dataSource.initDB();
+    dataSourceTest.initDB();
     // normal close
-    dataSource.closeDB();
+    dataSourceTest.closeDB();
     // closing already closed db.
-    dataSource.closeDB();
+    dataSourceTest.closeDB();
     // closing again to make sure the lock is free. If not test will hang.
-    dataSource.closeDB();
+    dataSourceTest.closeDB();
 
-    assertFalse("Database is still alive after closing.", dataSource.isAlive());
+    assertFalse("Database is still alive after closing.", dataSourceTest.isAlive());
   }
-
 }
