--- src/test/java/org/tron/storage/leveldb/LevelDbDataSourceImplTest.java
@@ -1,130 +1,0 @@
-/*
- * Copyright (c) [2016] [ <ether.camp> ]
- * This file is part of the ethereumJ library.
- *
- * The ethereumJ library is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * The ethereumJ library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with the ethereumJ library. If not, see <http://www.gnu.org/licenses/>.
- */
-
-package org.tron.storage.leveldb;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-
-import java.util.HashMap;
-import java.util.Map;
-import org.junit.Before;
-import org.junit.Test;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.tron.common.storage.leveldb.LevelDbDataSourceImpl;
-import org.tron.common.utils.ByteArray;
-import org.tron.core.Constant;
-import org.tron.core.config.Configuration;
-import org.tron.core.config.args.Args;
-
-public class LevelDbDataSourceImplTest {
-
-  private static final Logger logger = LoggerFactory.getLogger("Test");
-
-  LevelDbDataSourceImpl dataSourceTest;
-
-  @Before
-  public void initDb() {
-    Args.setParam(new String[]{}, Configuration.getByPath(Constant.TEST_CONF));
-    dataSourceTest = new LevelDbDataSourceImpl(Constant.OUTPUT_DIR, "test");
-  }
-
-  @Test
-  public void testPutGet() {
-    //dataSourceTest.resetDb();
-    String key1 = "2c0937534dd1b3832d05d865e8e6f2bf23218300b33a992740d45ccab7d4f519";
-    byte[] key = key1.getBytes();
-    dataSourceTest.initDB();
-    String value1 = "50000";
-    byte[] value = value1.getBytes();
-
-    dataSourceTest.putData(key, value);
-
-    assertNotNull(dataSourceTest.getData(key));
-    assertEquals(1, dataSourceTest.allKeys().size());
-    assertEquals("50000", ByteArray.toStr(dataSourceTest.getData(key1.getBytes())));
-
-  }
-
-  @Test
-  public void testRest() {
-    LevelDbDataSourceImpl dataSource = new LevelDbDataSourceImpl(Constant.OUTPUT_DIR, "test2");
-    dataSource.resetDb();
-    assertEquals(0, dataSource.allKeys().size());
-    dataSource.closeDB();
-  }
-
-  @Test
-  public void testupdateByBatchInner() {
-    LevelDbDataSourceImpl dataSource = new LevelDbDataSourceImpl(Constant.OUTPUT_DIR, "test3");
-    dataSource.initDB();
-    String key1 = "431cd8c8d5abe5cb5944b0889b32482d85772fbb98987b10fbb7f17110757350";
-    String value1 = "50000";
-    String key2 = "431cd8c8d5abe5cb5944b0889b32482d85772fbb98987b10fbb7f17110757351";
-    String value2 = "10000";
-
-    Map<byte[], byte[]> rows = new HashMap<>();
-    rows.put(key1.getBytes(), value1.getBytes());
-    rows.put(key2.getBytes(), value2.getBytes());
-
-    dataSource.updateByBatch(rows);
-
-    assertEquals("50000", ByteArray.toStr(dataSource.getData(key1.getBytes())));
-    assertEquals("10000", ByteArray.toStr(dataSource.getData(key2.getBytes())));
-    assertEquals(2, dataSource.allKeys().size());
-  }
-
-  @Test
-  public void testdeleteData() {
-    LevelDbDataSourceImpl dataSource = new LevelDbDataSourceImpl(Constant.OUTPUT_DIR, "test4");
-    dataSource.initDB();
-    String key1 = "431cd8c8d5abe5cb5944b0889b32482d85772fbb98987b10fbb7f17110757350";
-    byte[] key = key1.getBytes();
-    dataSource.deleteData(key);
-    byte[] value = dataSource.getData(key);
-    String s = ByteArray.toStr(value);
-    assertNull(s);
-
-  }
-
-  @Test
-  public void testallKeys() {
-    LevelDbDataSourceImpl dataSource = new LevelDbDataSourceImpl(Constant.OUTPUT_DIR, "test5");
-    dataSource.initDB();
-    dataSource.resetDb();
-
-    String key1 = "431cd8c8d5abe5cb5944b0889b32482d85772fbb98987b10fbb7f17110757321";
-    byte[] key = key1.getBytes();
-
-    String value1 = "50000";
-    byte[] value = value1.getBytes();
-
-    dataSource.putData(key, value);
-    String key3 = "431cd8c8d5abe5cb5944b0889b32482d85772fbb98987b10fbb7f17110757091";
-    byte[] key2 = key3.getBytes();
-
-    String value3 = "30000";
-    byte[] value2 = value3.getBytes();
-
-    dataSource.putData(key2, value2);
-    assertEquals(2, dataSource.allKeys().size());
-    dataSource.resetDb();
-  }
-}
